features/webpush/flask/pyproject.toml
[project]
name = "flask-webpush-template"
version = "0.1.0"
description = "Flask WebPush PWA template for local network deployments"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.0.0",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pywebpush>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
